+++ internal/cfd/linalg/jacobi.go
@@ -2,19 +2,19 @@ package linalg
 
 type Jacobi struct {
 	maxIterations int
-	tolerance     float32
+	tolerance     float64
 }
 
-func (j *Jacobi) Solve(sys *System, x []float32) []float32 {
+func (j *Jacobi) Solve(sys *System, x []float64) []float64 {
 	matrix := sys.A
 	rhs := sys.B
 
 	n := len(rhs)
-	xNew := make([]float32, n) // this is always going to be rubbish
+	xNew := make([]float64, n) // this is always going to be rubbish
 
 	for iter := 0; iter < j.maxIterations; iter++ {
 		for i := range n {
-			var sum float32 = 0.0
+			var sum float64 = 0.0
 
 			for k := range n {
 				if i != k {
@@ -33,10 +33,10 @@ func (j *Jacobi) Solve(sys *System, x []float32) []float32 {
 
 type JacobiDefinition struct {
 	maxIterations int
-	tolerance     float32
+	tolerance     float64
 }
 
-func NewJacobi(maxIterations int, tolerance float32) SolverDefinition {
+func NewJacobi(maxIterations int, tolerance float64) SolverDefinition {
 	return &JacobiDefinition{
 		maxIterations: maxIterations,
 		tolerance:     tolerance,
