+++ internal/cfd/linalg/steepestdesc.go
@@ -2,12 +2,12 @@ package linalg
 
 type steepestDescent struct {
 	maxIterations int
-	tolerance     float32
-	residual      []float32
-	matr          []float32
+	tolerance     float64
+	residual      []float64
+	matr          []float64
 }
 
-func (sd *steepestDescent) Solve(sys *System, x []float32) []float32 {
+func (sd *steepestDescent) Solve(sys *System, x []float64) []float64 {
 	matrix := sys.A
 	rhs := sys.B
 	r := sd.residual
@@ -18,7 +18,7 @@ func (sd *steepestDescent) Solve(sys *System, x []float32) []float32 {
 		r[i] = rhs[i] - val
 	}
 
-	var rDotr float32 = 0
+	var rDotr float64 = 0
 	for _, val := range r {
 		rDotr += val * val
 	}
@@ -29,7 +29,7 @@ func (sd *steepestDescent) Solve(sys *System, x []float32) []float32 {
 	for iter := 0; iter < sd.maxIterations && rDotr > threshold; iter++ {
 		Ar = matrix.MatVec(r, Ar)
 
-		var rDotAr float32 = 0
+		var rDotAr float64 = 0
 		for i, val := range r {
 			rDotAr += val * Ar[i]
 		}
@@ -62,10 +62,10 @@ func (sd *steepestDescent) Solve(sys *System, x []float32) []float32 {
 
 type steepestDescentDefinition struct {
 	maxIterations int
-	tolerance     float32
+	tolerance     float64
 }
 
-func NewSteepestDescent(maxIterations int, tolerance float32) SolverDefinition {
+func NewSteepestDescent(maxIterations int, tolerance float64) SolverDefinition {
 	return &steepestDescentDefinition{
 		maxIterations: maxIterations,
 		tolerance:     tolerance,
@@ -76,7 +76,7 @@ func (sdd *steepestDescentDefinition) Resolve(n int) Solver {
 	return &steepestDescent{
 		maxIterations: sdd.maxIterations,
 		tolerance:     sdd.tolerance,
-		residual:      make([]float32, n),
-		matr:          make([]float32, n),
+		residual:      make([]float64, n),
+		matr:          make([]float64, n),
 	}
 }
